# docs/site/static/openapi.yaml
openapi: 3.0.3
info:
  title: 区块链实验室网站后端API
  description: |
    区块链实验室网站后端接口文档（基于控制器代码校准）。
    
    鉴权说明：
    - 当前安全配置为 `permitAll`，技术上所有接口可访问。
    - 部分写接口依赖 Session（Cookie: JSESSIONID）或约定“隐藏入口访问”，文档中以“业务权限”标注。
  version: 1.0.0
  contact:
    email: support@blockchain-lab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082/api
    description: 开发环境
  - url: http://test.blockchain-lab.com:8082/api
    description: 测试环境
  - url: http://pre.blockchain-lab.com:8082/api
    description: 预发环境
  - url: http://blockchain-lab.com:8082/api
    description: 生产环境
...
paths:
...
  /posts:
    get:
      tags:
        - 博客文章管理
      summary: 获取所有文章
      description: 分页获取所有已发布的博客文章
      operationId: getAllPublishedPosts
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 每页大小
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: 排序字段
          schema:
            type: string
            default: createdAt
        - name: sortDirection
          in: query
          description: 排序方向
          schema:
            type: string
            default: DESC
            enum: [ASC, DESC]
      responses:
        '200':
          description: 文章列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePostDto'
    post:
      tags:
        - 博客文章管理
      summary: 创建文章
      description: 创建新的博客文章（隐藏入口访问）
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDto'
      responses:
        '200':
          description: 创建的文章信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                example1:
                  summary: 未登录
                  value:
                    message: 未找到有效 Session，浏览器未携带 JSESSIONID 或 Session 已失效
...
  /posts/latest:
    get:
      tags:
        - 博客文章管理
      summary: 获取最新文章
      description: 获取最新的博客文章
      operationId: getLatestPosts
      parameters:
        - name: limit
          in: query
          description: 文章数量
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: 最新文章列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
  /posts/{id}:
    get:
      tags:
        - 博客文章管理
      summary: 根据ID获取文章
      description: 根据ID获取博客文章详细信息
      operationId: getPostById
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
      responses:
        '200':
          description: 文章详细信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
    put:
      tags:
        - 博客文章管理
      summary: 更新文章
      description: 更新博客文章（隐藏入口访问）
      operationId: updatePost
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDto'
      responses:
        '200':
          description: 更新后的文章信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
    delete:
      tags:
        - 博客文章管理
      summary: 删除文章
      description: 删除博客文章（隐藏入口访问）
      operationId: deletePost
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
  /posts/{id}/status:
    patch:
      tags:
        - 博客文章管理
      summary: 更新文章状态
      description: 更新博客文章状态（隐藏入口访问）
      operationId: updatePostStatus
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
        - name: status
          in: query
          required: true
          description: 文章状态
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
      responses:
        '200':
          description: 更新后的文章信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
  /posts/{id}/like:
    post:
      tags:
        - 博客文章管理
      summary: 点赞文章
      description: 点赞博客文章
      operationId: likePost
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
      responses:
        '200':
          description: 点赞成功
    delete:
      tags:
        - 博客文章管理
      summary: 取消点赞文章
      description: 取消点赞博客文章
      operationId: unlikePost
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
      responses:
        '200':
          description: 取消点赞成功
  /posts/{id}/display-order:
    patch:
      tags:
        - 博客文章管理
      summary: 更新文章显示顺序
      description: 更新博客文章显示顺序（隐藏入口访问）
      operationId: updatePostDisplayOrder
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: integer
        - name: displayOrder
          in: query
          required: true
          description: 显示顺序
          schema:
            type: integer
      responses:
        '200':
          description: 更新成功
...
  /posts/search:
    get:
      tags:
        - 博客文章管理
      summary: 搜索文章
      operationId: searchPosts
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByCreatedAtParam'
        - $ref: '#/components/parameters/SortDirectionParam'
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePostDto'
  /posts/tag/{tag}:
    get:
      tags:
        - 博客文章管理
      summary: 根据标签获取文章
      operationId: getPostsByTag
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 文章列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
...
components:
  schemas:
...
    PostDto:
      type: object
      required:
        - title
        - content
      properties:
        id:
          type: integer
          description: 文章ID
        title:
          type: string
          description: 文章标题
        content:
          type: string
          description: 文章内容
        summary:
          type: string
          description: 文章摘要
        author:
          type: string
          description: 作者
        tags:
          type: string
          maxLength: 500
          description: 标签（逗号分隔）
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          description: 状态
        viewCount:
          type: integer
          description: 浏览量
        likeCount:
          type: integer
          description: 点赞数
        commentCount:
          type: integer
          description: 评论数
        allowComments:
          type: boolean
          description: 是否允许评论
        featured:
          type: boolean
          description: 是否特色
        publishedAt:
          type: string
          format: date-time
          description: 发布时间
        displayOrder:
          type: integer
          description: 显示顺序
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
...
    PagePostDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PostDto'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          $ref: '#/components/schemas/Sort'
        offset:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: 错误码
        message:
          type: string
          description: 错误信息
        data:
          type: object
          nullable: true
          description: 数据
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从0开始）
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: 每页大小
      schema:
        type: integer
        default: 20
        minimum: 1
...
    SortByCreatedAtParam:
      name: sortBy
      in: query
      description: 排序字段
      schema:
        type: string
        default: createdAt
...
    SortDirectionParam:
      name: sortDirection
      in: query
      description: 排序方向
      schema:
        type: string
        default: DESC
        enum: [ASC, DESC]
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session认证
security:
  - sessionAuth: []